Optimised NSFW parameter handling.

+++ src/commands/fun/info/poll.js
@@ -1,9 +1,28 @@
 const PREFIX = process.env.PREFIX;
+const fs = require('fs');
 
 module.exports = {
     run: async(client, message, args) => {
+        if (message.content == PREFIX) return;
         message.delete();
         if(message.content == `${PREFIX}poll`) return;
+        let bannedWords = fs.readFileSync('./events/bannedwords.txt').toString().split("\r\n");
+        let bannedPhrases = fs.readFileSync('./events/bannedphrases.txt').toString().split("\r\n");
+        let msg = message.content.toLowerCase();
+        let wordsOnlyMsg = msg.replace(/[.?!#$%^&*,-_+=]/g, ' ');
+        let words = wordsOnlyMsg.split(/\s+/);
+
+        // Checks if parameter is an nsfw term. Blocks command in non-nsfw channels.
+        if (!message.channel.nsfw) {
+            for (let i = 0; i < bannedWords.length; i++) { 	
+                if (words.includes(bannedWords[i])) return;
+            }
+
+            for (let j = 0; j < bannedPhrases.length; j++) { 	
+                if (msg.includes(bannedPhrases[j])) return;
+            }
+        }
+
         let pollEmbed = {
             color: `RANDOM`,
             title: `${message.author.username} asks...`,
